Remove Mark IV properties that cannot really be set.
In particular, ntrack is defined by the dimension of words, so
cannot be changed.  And once ntrack is defined, the header size,
framesize, and payloadsize are all known.

The only use of setting ntrack before was that it filled in
defaults for headstack_id and track_id; that is now done directly
in ``fromvalues``.